Make the settings dialog slightly more generic.
git-svn-id: https://kraigsandroid.googlecode.com/svn/trunk@76 81ac7b4a-400a-11de-9cfd-b5a970547080

+++ android/alarmclock/src/com/angrydoughnuts/android/alarmclock/AlarmSettings.java
@@ -6,7 +6,7 @@ import android.os.Parcel;
 import android.os.Parcelable;
 
 public class AlarmSettings implements Parcelable {
-  static public final int DEFAULT_SETTINGS_ID = -1;
+  static public final long DEFAULT_SETTINGS_ID = -1;
 
   private Uri tone;
